**lxml** was written as **lxlm** that was confusing.
The lxml XML toolkit is a Pythonic binding for the C libraries libxml2 and libxslt. It is unique in that it combines the speed and XML feature completeness of these libraries with the simplicity of a native Python API, mostly compatible but superior to the well-known ElementTree API. The latest release works with all CPython versions from 2.6 to 3.4. See the introduction for more information about background and goals of the lxml project.
Source: www.lxml.de